Use better name for fetching rate data
