+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2020-09-14 16:02:48 UTC using RuboCop version 0.90.0.
+# on 2020-09-14 18:17:10 UTC using RuboCop version 0.90.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -12,14 +12,6 @@
 Gemspec/RequiredRubyVersion:
   Exclude:
     - 'money-historical-bank.gemspec'
-
-# Offense count: 1
-# Cop supports --auto-correct.
-# Configuration parameters: AllowAliasSyntax, AllowedMethods.
-# AllowedMethods: alias_method, public, protected, private
-Layout/EmptyLinesAroundAttributeAccessor:
-  Exclude:
-    - 'lib/money/bank/historical_bank.rb'
 
 # Offense count: 3
 # Configuration parameters: AllowSafeAssignment.
@@ -88,32 +80,19 @@ Security/MarshalLoad:
   Exclude:
     - 'lib/money/bank/historical_bank.rb'
 
-# Offense count: 1
-Security/Open:
-  Exclude:
-    - 'lib/money/bank/open_exchange_rates_loader.rb'
-
 # Offense count: 2
 Style/Documentation:
   Exclude:
     - 'spec/**/*'
     - 'test/**/*'
     - 'lib/money/bank/historical_bank.rb'
-    - 'lib/money/bank/open_exchange_rates_loader.rb'
+    - 'lib/open_exchange_rates_fetcher.rb'
 
 # Offense count: 1
 # Configuration parameters: MinBodyLength.
 Style/GuardClause:
   Exclude:
     - 'lib/money/bank/historical_bank.rb'
-
-# Offense count: 4
-# Cop supports --auto-correct.
-Style/StringConcatenation:
-  Exclude:
-    - 'lib/money/bank/historical_bank.rb'
-    - 'lib/money/bank/open_exchange_rates_loader.rb'
-    - 'test/historical_bank_test.rb'
 
 # Offense count: 3
 # Cop supports --auto-correct.
